Refactor files to have a rule based generator and model based generator
