+++ generator/src/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['generator', 'model', 'data_processing', 'trainer', 'transformations', 'utils', 'visualizer']
+__all__ = ['generator', 'rule_based_generator', 'model', 'data_processing', 'trainer', 'transformations', 'utils', 'visualizer']
