+++ generator/src/model/model.py
@@ -20,7 +20,7 @@ class ModelArgs:
     n_heads: int = 8
     multiple_of: int = 64  # make SwiGLU hidden layer size multiple of large power of 2
     norm_eps: float = 1e-5
-    rope_theta: float = 1000
+    rope_theta: float = 100
 
     max_batch_size: int = 32
     max_seq_len: int = 258
