+++ generator/src/trainer/trainer.py
@@ -213,11 +213,11 @@ class AutoEncoderTrainer(Trainer):
 if __name__ == "__main__":
     # EXAMPLE CODE FOR TRANSFORMER TRAINING
     trainer = Trainer(
-        n_epochs=20,
-        lr=1e-4,
+        n_epochs=100,
+        lr=2e-3,
         batch_size=4,
-        batch_norm_momentum=0.1,
-        extra_augmentation=lambda image: transformations.transformations_for_training(image, crop_size=32)
+        batch_norm_momentum=0.01,
+        extra_augmentation=lambda image: transformations.transformations_for_training(image, crop_size=16)
     )
     args = model.ModelArgs()
     model = model.SpatioTemporalTransformer(args).to(DEVICE)
@@ -226,14 +226,16 @@ if __name__ == "__main__":
     # get the first batch of the loader
     train_loader, test_loader = data_processing.get_dataloader(
         batch_size=1,
-        transform=lambda image: transformations.transformations_for_evaluation(image, crop_size=32)
+        transform=lambda image: transformations.transformations_for_evaluation(image, crop_size=16)
     )
 
     model.eval()
     batch = next(iter(test_loader)).to(DEVICE)
     predictions = model(batch[:, :-1])
     predictions_unnormalized = transformations.unnormalize_image(predictions)
-    visualizer.visualize_tensor_image(predictions_unnormalized[0][1])
+    batch_unnormalized = transformations.unnormalize_image(batch)
+    visualizer.visualize_tensor_image(predictions_unnormalized[0][45])
+    visualizer.visualize_tensor_image(batch_unnormalized[0][46])
 
     # EXAMPLE CODE FOR AUTOENCODER TRAINING
     # autoencoder_trainer = AutoEncoderTrainer(
